+++ CHANGES.md
@@ -5,9 +5,12 @@ These are the changes brought in each versions:
 v0.0.2-alpha (TBD)
 =========================
 
+Non-breaking changes:
+---------------------
+* New `--date-report` argument
+
 Other changes:
 --------------
-
 * Changes to the development toolchain and test suite
 
 v0.0.1-alpha (2020-09-10)